rearrange spinner
